requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Search and web scraping
ddgs
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# JSON schema validation
jsonschema>=4.19.0

# Data processing and utilities
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# HTTP client improvements
httpx>=0.24.0
aiohttp>=3.8.0

# Development and debugging
python-dateutil>=2.8.0
typing-extensions>=4.5.0

# Optional: For better error handling
tenacity>=8.2.0

# Optional: For advanced text processing
tiktoken>=0.5.0

# Core dependencies
pydantic>=2.5.0
jsonschema>=4.17.0
requests>=2.31.0
Jinja2>=3.1.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
mypy>=1.7.0